+++ src/app/pages/Governance/routes.ts
@@ -5,6 +5,7 @@ export enum GovernancePathQueries {
 }
 
 export const GovernanceSubPages = {
+  newProposal: `${AppPages.GovernancePage}/newProposal`,
   proposals: `${AppPages.GovernancePage}/proposals`,
   proposal: `${AppPages.GovernancePage}/proposals/:${GovernancePathQueries.proposalIndex}`,
 };
