+++ src/app/pages/Governance/slice.ts
@@ -17,7 +17,6 @@ export const initialState: ContainerState = {
   pairSearchInput: "",
   selectedPoolProviders: [],
   isVotingForFarms: false,
-  isNewProposalFormOpen: true,
 };
 
 const governancePageSlice = createSlice({
@@ -73,9 +72,6 @@ const governancePageSlice = createSlice({
     voteForFarms: (state, action: PayloadAction<void>) => {},
     setIsVotingForFarms: (state, action: PayloadAction<boolean>) => {
       state.isVotingForFarms = action.payload;
-    },
-    setIsNewProposalFormOpen(state, action: PayloadAction<boolean>) {
-      state.isNewProposalFormOpen = action.payload;
     },
   },
 });
