+++ src/app/pages/Governance/subPages/proposals/index.tsx
@@ -3,7 +3,6 @@ import { styled } from "@mui/material";
 import { Max1040 } from "app/components/wrappers/max1040";
 import { VotePower } from "../../components/votePower";
 import { TopWrapper } from "./components/topWrapper";
-import { NewProposalForm } from "./newProposalForm";
 import { mobile } from "styles/media";
 import { useStakingPageSlice } from "app/pages/Staking/slice";
 
@@ -12,7 +11,6 @@ export const Proposals: FC = () => {
 
   return (
     <StyledMax1040 m="auto" mt={2}>
-      <NewProposalForm />
       <Header>
         <VotePower />
       </Header>
